refactoring services to match to a String ID no more Long id!!
