+++ Server/src/main/java/ch/amphytrion/project/services/OpeningHourService.java
@@ -1,6 +1,6 @@
 package ch.amphytrion.project.services;
 
-import ch.amphytrion.project.entities.OpeningHour;
+import ch.amphytrion.project.entities.databaseentities.OpeningHour;
 import ch.amphytrion.project.repositories.OpeningHourRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -28,7 +28,7 @@ public class OpeningHourService implements IGenericService<OpeningHour> {
     }
 
     @Override
-    public OpeningHour findById(long id) {
+    public OpeningHour findById(String id) {
         try {
             return openingHourRepository.findById(id).orElseThrow(Exception::new);
         } catch (Exception e) {
@@ -43,7 +43,7 @@ public class OpeningHourService implements IGenericService<OpeningHour> {
     }
 
     @Override
-    public void deleteById(long id) {
+    public void deleteById(String id) {
         openingHourRepository.deleteById(id);
     }
