+++ Server/src/main/java/ch/amphytrion/project/services/TagService.java
@@ -1,6 +1,6 @@
 package ch.amphytrion.project.services;
 
-import ch.amphytrion.project.entities.Tag;
+import ch.amphytrion.project.entities.databaseentities.Tag;
 import ch.amphytrion.project.repositories.TagRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -28,7 +28,7 @@ public class TagService implements IGenericService<Tag> {
     }
 
     @Override
-    public Tag findById(long id) {
+    public Tag findById(String id) {
         try {
             return tagRepository.findById(id).orElseThrow(Exception::new);
         } catch (Exception e) {
@@ -43,7 +43,7 @@ public class TagService implements IGenericService<Tag> {
     }
 
     @Override
-    public void deleteById(long id) {
+    public void deleteById(String id) {
         tagRepository.deleteById(id);
     }
